feat: support SEP-07 callbacks
using existing multisig code as much as possible, so
when wallet recieves SEP-07 requests with callback it stores a new signatureRequest
to activate multisig flow (even if there is just one signature).

When callback is provided the wallet won't publish the tx to the network (according to SEP07 spec).

Different cases handled by different parts of the code (Solar's design):
1. opening web+stellar:tx link by click or QR code scan
2. opening web+stellar:pay link by click or QR code scan
3. opening web+stellar:pay link from Payment Dialog using camera scanner