fix: allow setting empty SEP-07 /pay fields
SEP-07 requires only 'destination' to be always present on the link, everything else is optional.
Let's allow user to enter amount, asset or memo should one or many of these aren't present in SEP-07 link.

If asset, amount or memo is present do not allow to change it — we treat SEP-07 /pay as a strict payment request